requirements.txt
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.51.0
